requirements.txt
opencv-python
mediapipe
PyQt6
SpeechRecognition
PyAutoGUI
numpy==1.24.4
pyaudio
pyttsx3
protobuf
pandas
scipy==1.11.3
tensorflow==2.14.0
playsound
spacy
scikit-learn
spacy-lookups-data
torch
openai-whisper
scipy 
psycopg2-binary
TTS==0.22.0
transformers==4.33.3
pynput
langchain==0.0.172 
chromadb 
wikipedia 
sentence-transformers==2.2.2
huggingface-hub==0.19.4
ollama
langchain-community==0.0.38
